Added main page component and refactored structure
Added the React component for the main page

Refactored the page by putting the nav bar and the popups into their own components

Refactored the scss files of the changed components.

Moved the (registration) folder into the (main) folder so they use the same layout file that uses the navbar component.