Various updates
Added logic to login and register to provide feedback to the user while logging in/registering

The user cannot use the back arrow button on browsers to go back to login after successfully logging in now.

The firebase database will now store data on quiz attempts and time taken in seconds.

Fixed #21